@@ -197,6 +197,12 @@ function BreakLineComponent({ style }: { style: CSSProperties }) {
   )
 }
 
+interface TokenRowProps {
+  data: Array<Currency | BreakLine>
+  index: number
+  style: CSSProperties
+}
+
 const formatAnalyticsEventProperties = (
   token: Token,
   index: number,
@@ -253,7 +259,7 @@ export default function CurrencyList({
   }, [currencies, otherListTokens])
 
   const Row = useCallback(
-    function TokenRow({ data, index, style }) {
+    function TokenRow({ data, index, style }: TokenRowProps) {
       const row: Currency | BreakLine = data[index]
 
       if (isBreakLine(row)) {
