@@ -4,7 +4,7 @@ import { useCallback, useEffect, useState } from 'react'
 export default function useCopyClipboard(timeout = 500): [boolean, (toCopy: string) => void] {
   const [isCopied, setIsCopied] = useState(false)
 
-  const staticCopy = useCallback((text) => {
+  const staticCopy = useCallback((text: string) => {
     const didCopy = copy(text)
     setIsCopied(didCopy)
   }, [])
